+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/authentication/LogoutAll.kt
@@ -1,8 +1,10 @@
 package net.folivo.trixnity.clientserverapi.model.authentication
 
+import io.ktor.http.*
 import io.ktor.resources.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
 import net.folivo.trixnity.core.HttpMethod
 import net.folivo.trixnity.core.HttpMethodType.POST
 import net.folivo.trixnity.core.MatrixEndpoint
@@ -16,4 +18,7 @@ import net.folivo.trixnity.core.model.UserId
 @HttpMethod(POST)
 data class LogoutAll(
     @SerialName("user_id") val asUserId: UserId? = null,
-) : MatrixEndpoint<Unit, Unit>
+) : MatrixEndpoint<Unit, Unit> {
+    @Transient
+    override val requestContentType: ContentType? = null
+}